+++ osmnx/stats.py
@@ -5,18 +5,17 @@
 # Web: https://github.com/gboeing/osmnx
 ################################################################################
 
-from __future__ import division
-import time
 import networkx as nx
 import numpy as np
 import pandas as pd
+import time
 
 from .simplify import clean_intersections
-from .utils_graph import get_largest_component
+from .utils import log
+from .utils_geo import euclidean_dist_vec
 from .utils_geo import great_circle_vec
 from .utils_graph import count_streets_per_node
-from .utils_geo import euclidean_dist_vec
-from .utils import log
+from .utils_graph import get_largest_component
